cas-svn-reporter/reporter/requirements.txt
astroid==3.2.4
certifi==2024.7.4
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dill==0.3.8
et-xmlfile==1.1.0
idna==3.8
isort==5.13.2
itsdangerous==2.2.0
jdcal==1.4.1
Jinja2==3.1.4
lazy-object-proxy==1.10.0
MarkupSafe==2.1.5
mccabe==0.7.0
mdutils==1.6.0
multi_key_dict==2.0.3
openpyxl==3.1.5
pbr==6.1.0
platformdirs==4.2.2
pylint==3.2.6
python-jenkins==1.8.2
requests==2.32.3
setuptools==74.0.0
six==1.16.0
toml==0.10.2
tomlkit==0.13.2
urllib3==2.2.2
Werkzeug==3.0.4
wrapt==1.16.0
